Refactor credit and employee services to filter by company ID; update user profile to include billing currency; add travel request functionality with CRUD operations.

+++ src/main/java/com/TravelMedicineAdvisory/Server/domain/credit/CreditRequest.java
@@ -3,10 +3,10 @@ package com.TravelMedicineAdvisory.Server.domain.credit;
 import java.math.BigDecimal;
 
 public record CreditRequest(
-    BigDecimal amount,
+    Integer amount,
     String type,
     String reference,
-    BigDecimal balanceAfter,
+    Integer balanceAfter,
     Long companyId,
     Long userId
 ) {}
